third_party/blink/web_tests/platform/generic/fast/filesystem/simple-readonly-file-object-expected.txt

Test readonly attributes of File.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

webkitRequestFileSystem(TEMPORARY, 100, fileSystemCallback, errorCallback);
removeAllInDirectory(fileSystem.root, createTestFile, errorCallback);
fileSystem.root.getFile(testFileName, {create:true}, getFileFromEntry, errorCallback);
testFileEntry.file(fileCallback, errorCallback);
trying to set readonly property testFile.size
testFile.size = 1
PASS testFile.size is still 0
trying to set readonly property testFile.type
testFile.type = 'application/octet-stream'
PASS testFile.type is still text/plain
trying to set readonly property testFile.name
testFile.name = 'bar'
PASS testFile.name is still testFileEntry.txt
PASS successfullyParsed is true

TEST COMPLETE
